pyproject.toml
[tool.poetry]
name = "redundanet"
version = "2.0.0"
description = "Distributed encrypted storage on a mesh VPN network"
authors = ["Alessandro De Filippo <alessandro@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/alessandrodefilippo/redundanet"
repository = "https://github.com/alessandrodefilippo/redundanet"
documentation = "https://redundanet.readthedocs.io"
keywords = ["distributed-storage", "vpn", "encryption", "mesh-network", "tahoe-lafs"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Distributed Computing",
    "Topic :: System :: Networking",
    "Topic :: Security :: Cryptography",
]
packages = [{include = "redundanet", from = "src"}]

[tool.poetry.scripts]
redundanet = "redundanet.cli.main:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
pydantic = "^2.5"
pydantic-settings = "^2.1"
pyyaml = "^6.0"
python-gnupg = "^0.5"
httpx = "^0.26"
rich = "^13.7"
structlog = "^24.1"
jinja2 = "^3.1"
jsonschema = "^4.21"
gitpython = "^3.1"
watchdog = "^4.0"
psutil = "^5.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-cov = "^4.1"
pytest-asyncio = "^0.23"
pytest-mock = "^3.12"
mypy = "^1.8"
ruff = "^0.1"
pre-commit = "^3.6"
types-PyYAML = "^6.0"
types-jsonschema = "^4.21"
hypothesis = "^6.92"
faker = "^22.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5"
mkdocs-material = "^9.5"
mkdocstrings = {extras = ["python"], version = "^0.24"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "TCH001", # Move application import into type-checking block (breaks runtime)
    "TCH002", # Move third-party import into type-checking block (breaks runtime)
    "TCH003", # Move stdlib import into type-checking block (breaks runtime)
    "UP007",  # Use X | Y for type annotations (requires Python 3.10+)
]

[tool.ruff.per-file-ignores]
"tests/**/*.py" = ["ARG001", "ARG002", "S101", "PTH123"]
"src/redundanet/cli/**/*.py" = ["ARG001"]  # Typer uses unused args for dependency injection

[tool.ruff.isort]
known-first-party = ["redundanet"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "gnupg.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "-ra",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = ["src/redundanet"]
branch = true
omit = ["*/tests/*", "*/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
